fix : main페이지 경로 root로 변경 (#22)
